# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Airport Transfer Aggregator API
  description: |
    REST API for aggregating airport transfer services from multiple suppliers.

    ## Features
    - Search across multiple suppliers in parallel
    - Unified booking interface
    - Signed, stateless tokens for offers and bookings
    - Idempotency support for booking operations

    ## Authentication
    No authentication required for API calls. Supplier credentials are managed server-side.

    ## Rate Limits
    - Search: 80 requests/minute
    - Booking: 50 requests/minute
  version: 1.0.0
  contact:
    name: Arcube Engineering
    email: engineering@arcube.com
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/pricing:
    post:
      tags:
        - Pricing
      summary: Calculate total price for an offer with selected amenities
      operationId: calculatePrice
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PricingRequest'
      responses:
        '200':
          description: Pricing calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer/Search not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Offer expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Pricing
      summary: Get price with amenities via query params
      description: Mozio-aligned GET support.
      operationId: getPriceWithAmenities
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema:
            type: string
        - name: searchId
          in: query
          required: true
          schema:
            type: string
        - name: offerId
          in: query
          required: true
          schema:
            type: string
        - name: amenities
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Pricing calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer/Search not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Offer expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pricing/{offerId}/amenities:
    get:
      tags:
        - Pricing
      summary: Get list of available amenities for an offer
      operationId: getAvailableAmenities
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema:
            type: string
        - name: offerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Amenities returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AmenityInfo'
        '400':
          description: Invalid offerId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Offer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Offer expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/alerts/dlq/simulate:
    post:
      tags:
        - AdminAlerts
      summary: Simulate adding items to DLQ
      operationId: simulateDlqItems
      parameters:
        - name: count
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
      responses:
        '200':
          description: DLQ items simulated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DlqSimulateResponse'
...
components:
  schemas:
...
    Money:
      type: object
      properties:
        value:
          type: number
        currency:
          type: string
        display:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        suggestedAction:
          type: string
          description: Suggested action (e.g., RETRY, RE_SEARCH)
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    PricingRequest:
      type: object
      required:
        - searchId
        - offerId
      properties:
        searchId:
          type: string
        offerId:
          type: string
          description: Signed offer ID (from search results)
        amenities:
          type: array
          items:
            type: string
          description: Selected amenity keys
    PricingResponse:
      type: object
      properties:
        searchId:
          type: string
        offerId:
          type: string
        basePrice:
          $ref: '#/components/schemas/Money'
        selectedAmenities:
          type: array
          items:
            $ref: '#/components/schemas/SelectedAmenity'
        totalPrice:
          $ref: '#/components/schemas/Money'
        message:
          type: string
    SelectedAmenity:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        price:
          $ref: '#/components/schemas/Money'
    AmenityInfo:
      type: object
      description: Amenity metadata returned by PricingService#getAvailableAmenities
      properties:
        key:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          $ref: '#/components/schemas/Money'
...
    DlqSimulateResponse:
      type: object
      properties:
        added:
          type: integer
        newSize:
          type: integer
...